Make vipsutil easier to turn lite
